debian: Clear out /etc/resolv.conf at the end
LAT uses the host's /etc/resolv.conf for DNS lookups when
installing the rootfs. It does not clear out the host information
when it is finished, so clear it out when we are done.

Test Plan
PASS Build ISO
PASS Boot ISO
PASS Check for empty /etc/resolv.conf

Story: 2009965
Task: 45527

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I0e52191eb99d866fb22a7687bfebf291f17d66db