ifupdown-extra packaging fix
ifupdown-extra was recently added to the set of packages that StarlingX
compiles.  Previously StarlingX was using a pre-compiled package provided
by upstream. THe update that introduced ifupdown-extra was...

https://review.opendev.org/c/starlingx/integ/+/842614

When the change was made, the packaging changes defining the iso content
were not updated.  As a result, the upstream version of ifupdown-extra
continued to be used.

This update the the iso packaging to include the locally compiled
ifupdown-extra, rather than the upstream version.

Testing:

Build and install the iso.
Verify that the StarlingX modified version is present by looking for
the new function 'get_prefix_length'.   e.g.

   cat /usr/etc/network/if-up.d/20static-routes | grep get_prefix_length

Closes-Bug: 1974229
Signed-off-by: Scott Little <scott.little@windriver.com>
Change-Id: Iab3c83ae8373923cb0358c7324c9e13542f50327