debian: Use "--force-unsafe-io" during build-image
Turn on the '--force-unsafe-io' when running dpkg/apt. The way
this option works is that it disables the fsync() system call
when the package is unpacked. This speeds up apt/dpkg installation
when debootstrap is running that creates the image. It is not
turned on by default in the running system because if the power
is cut then it might result in data corruption.

This change only affects buidling ISO and is not turned on by default
on a running system.

Test Plan
Enable --force-unsafe-io during build
Run build-image succesfully

Story: 2009221
Task: 45757

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: If6e8169c6bd8a29c5dfee117190c3b6ba88f0801