Debian: Add the 'dig' utility to the LAT installer initramfs
The dns dig utility is needed to resolve the ipv4 vs ipv6
addressing mode for the interfaces creation section of the
kickstart.

This update adds the bind9-dnsutils, which contains the
dig utility, to the LAT initramfs.

Test Plan:

PASS: Verify Debian build, iso creation and system install
PASS: Verify dig utility is added to the LAt initramfs

Story: 2009968
Task: 45934
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>
Change-Id: I7774f2eb7dc9cabf36056617576b4a335dc39ef7