Ignore --all/lo* for ifupdown/nslcd scripts
It was detected the networking.service is marked as failing after reboot.
It happpens because "ifup -a" is executed by the service.
It starts to run the scripts in /etc/network/interfaces.d/.
But several scripts in ifupdown-extra are not prepared to handle "-all".

In the case of nss-pam-ldapd the script /etc/network/if-up.d/nslcd
is failing when there are loopback interfaces with label (lo:X) as the
script only tests the interface "lo".

Test Plan (Debian only - AIO-SX and AIO-DX)):
PASS  Check systemctl status networking.service after unlock

Partial-Bug: #1983503

Signed-off-by: Fabiano Mercer <fabiano.correamercer@windriver.com>

Depends-On: https://review.opendev.org/c/starlingx/integ/+/852049
Change-Id: I67a3a3113d13f811613c647d3176050417f5a909