aptly docker: change nginx log files to /var/aptly
The original log files of nginx are under '/var/log/nginx/'
which is not in local disk and it hard to collect these logs.
This commit changed these logs to the shared volume 'var/aptly'
of aptly docker.

Test Plan:
Pass: Create a new build environment and check with:
      $stx control enter --dockername repomgr
      $cat /var/aptly/nginx_error.log
      $cat /var/aptly/nginx_access.log
      exit from the repomgr:
      $cd ${STX_BUILD_HOME}
      $cat aptly/nginx_error.log
      $cat aptly/nginx_access.log
      nginx works and it logs into the above files

Story: 2008846
Task: 46647

Signed-off-by: Haiqing Bai <haiqing.bai@windriver.com>
Change-Id: Ibeeb7e1c2b98c3a5d8a78acd5bb269b45550589c