Update LAT container to support patches
This commit adds the necessary tooling and
env variables to the LAT container in order to
support patch creation.

The current patch process that has the
option to reuse initramfs and for that we
need access to the gpg keys used by LAT
to check in and sign the ostree commit,
as LAT already builds the image it makes
sense to also build the patches in the LAT
container.

Story: 2008862
Task: 46788

Signed-off-by: Luis Sampaio <luis.sampaio@windriver.com>
Change-Id: I1ae93bfe29e8003b64b65eb06d1d1ae65ef94067