Look for kickstarts in kickstart feed directory
LAT packages the kickstarts in the iso if it finds them in the feed
directory.

However, LAT is currently expecting to see them in the base
feed directory ; /var/www/pages/feed/rel-xx.xx

This is not where the kickstarts need to be. So the kickstarts
themselves are copying them to the feed's kickstart directory
at /var/www/pages/feed/rel-xx.xx/kickstart.

Along with the other kickstart feed cleanup in this update
bundle, see depends-on updates. This update moves the packaging
of the kickstarts into the kickstart directory and removes the
copy in the kickstart. This gets rid of the orphaned kickstarts
in the feed base dir in favor of now patchable kickstarts.

Test Plan:

PASS: Verify Debian build and AIO DX Install (usb and pxe).
PASS: Verify kickstarts in kickstart feed dir without copy.

Depends-On: https://review.opendev.org/c/starlingx/metal/+/864596

Story: 2009968
Task: 46836
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>
Change-Id: Ia393c2ce880d9e1336fdea8401ab8f934146bf4b