pkgbuilder: Stop chroot sessions before refreshing chroots
The build option 'refresh_chroots' will fail if one of the
chroot's flag is busy and the session is active.
This commit forces to stop all the sessions of chroots and
set the flag to idle to make refresh chroots works.
This commit requires rebuilding the pkgbuilder container.

Test Plan:
Pass: build-pkgs -a --parallel 10 --refresh_chroots
      build-pkgs -c -p <package> --refresh_chroots

Story: 2008846
Task: 46973

Signed-off-by: Haiqing Bai <haiqing.bai@windriver.com>
Change-Id: I4eff4c7d316e8c747d829460526f1f01258062bc