get_pod_name: Only return running pods
get_pod_name was observed to return a list of all
pods satisfying the query without consideration of
the state of the pod. In one observed scenario the name of
a runing pod and an terminating pod were returned.

All subsequent commands used the podname with out checking
for multiple entries.  They clearly expect only a single
name, not multiple.

Modify get_pod_name to only return a single rinning pod name.

Closes-bug: 2015885
Signed-off-by: Scott Little <scott.little@windriver.com>
Change-Id: I0b7bf22c4481fc4a81012fc14ac9d388afd6b065