Relocate pxeboot-update script to writable dir
Details: The pxeboot-update script is relocated to /etc/ in 23.09.
The reason is /usr/sbin/ is ostree managed directory. During the
upgrade, import.sh can't copy the pxeboot-update script to /usr/sbin/
anymore. The /etc/ is the new location for the pxeboot-update script.

Test Plan:

PASS: built the iso
PASS: upgraded from 22.12 to 23.09
PASS: downgrade from 23.09 to 22.12

Task: 47913
Story: 2010651
Change-Id: I5954b6700bc27a07b5de88934481613c06b87431
Signed-off-by: junfeng-li <junfeng.li@windriver.com>