Update puppet-memcached to R 8.2.0
Removing the puppet-module-saz-memcached release 3.1.0-1
since it doesn't allow the configuration of FQDN/hostname to
memcached service bind.

The puppet-memcached R 8.2.0 allows this configuration and will be
built from source.
It is necessary to configure the parameter "listen" instead of
"listen_ip" (deprecated). The use of FQDN is necessary to allow the
reconfiguration of the MGMT network.

Test plan ( Debian only )
 - AIO-SX and AIO-DX virtualbox installation
 - netstat -tupl ( no services are using the MGMT IP address )

Story: 2010722
Task: 48466
Depends-on: https://review.opendev.org/c/starlingx/integ/+/887534

Change-Id: Ib83a4de2c546e5b162d88619de81a980624e36b8
Signed-off-by: Fabiano Correa Mercer <fabiano.correamercer@windriver.com>