Create and populate openstack build layer
With the STX-Openstack upversion to Antelope, we need to update the
downloader list so the dependencies are fetched and used correctly.
For that, a new `openstack` layer is being created, with the [1]
change, this new layer will not override the packages defined in the
base-bullseye.lst, but instead will create a new aptly repo for it.
This way, both versions for these packages can coexist without affecting
each other.
Since these new versions are not debian/bullseye (oldstable), but
instead debian/trixie (testing), there was a need to add the direct
link to the Debian package (via snapshot.debian)

Note that this change will NOT impact the default StarlingX build, as
this change is on the `f/antelope` branch.

[1] https://review.opendev.org/c/starlingx/root/+/893075

Test Plan:
PASS: Download all dependencies (downloader -s -b -B std,rt)
PASS: Build all packages on openstack layer (build-pkgs -c -a -l
      openstack)

Depends-On: https://review.opendev.org/c/starlingx/root/+/896774

Story: 2010797
Task: 48694

Change-Id: I25fcb7fa5956a039d2a11bb13267867691ac33ce
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>
(cherry picked from commit b76a68227fe259d6586897ed365cffa22ea63927)