debbuilder: Support layer-specific repo in sbuild
In [1], a mechanism was proposed to create layer-specific aptly binary
repositories in addition to the existing `deb-local-binary`, to better
separate the binary dependencies of each layer.

For this new mechanism to work, new repositories must be dynamically
taken into account by `sbuild`, by adding a new `--extra-repository`
argument referring to the layer of the package being built, which is
precisely what this change does.

[1] https://review.opendev.org/c/starlingx/root/+/893095

Test Plan:
PASS - Build packages of different layers successfully

Story: 2010797
Task: 48698

Change-Id: Ib09ddf042584bc418c6a07773e920901961938ee
Signed-off-by: Luan Nunes Utimura <LuanNunes.Utimura@windriver.com>