Move OPAE tools package dependencies to flock
In [1], dependencies for the OPAE tools package (opae-sdk) were added in
the `containers` layer, despite the package itself being on the `flock`
layer (`starlingx/utilities`).

This isn't a problem when all downloaded binaries go to the same
repository: `deb-local-binary`. However, now that we are considering
creating one repository per layer ([2] and [3]), these dependencies need
to be on the appropriate layer. In this case, on the same layer as the
`opae-sdk`.

[1] https://review.opendev.org/c/starlingx/tools/+/862743
[2] https://review.opendev.org/c/starlingx/tools/+/893105
[3] https://review.opendev.org/c/starlingx/root/+/893095

Test Plan:
PASS - Run:
         $ repo_manage.py clean
       and verify that it cleaned all existing repositories:
         $ repo_manage.py list
PASS - Run `downloader` successfully and verify that the dependencies
       were uploaded to `deb-local-binary-flock` instead of
       `deb-local-binary-containers`
PASS - Build `opae-sdk` package
PASS - Build all packages
PASS - Build image

Story: 2010797
Task: 48714

Change-Id: Ic9a8663102872fdaf9de9e45aadbc78d328c8f4e
Signed-off-by: Luan Nunes Utimura <LuanNunes.Utimura@windriver.com>