Add bullseye Openstack packages to common layer
Using patchset 1 of [1], the build-image failed due to missing packages
that were removed with this change. After examining which packages were
required on that repo (ones that had patches that the platform uses), we
need to add the remaining packages to the base-bullseye.lst list, so
both the build of the python-openstackclient package, and the
build-image is successful.

Test Plan:
PASS: Run downloader -s -b -B std,rt
PASS: Build python-openstackclient
PASS: Build StarlingX ISO image

Depends-On: https://review.opendev.org/c/starlingx/manifest/+/893085

Story: 2010797
Task: 48692

Change-Id: I5368a3df88b4d30a9fb5d8479d830354b52cb091
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>