Additional Apt-ostree build tool fixes
LAT configures the ISO with the apt repositories from
the aptly repositories, which is the correct behaviour.
But it does not clean up after itself and leaves
the /etc/apt/sources.list as is when the build is complete.

Leaving the /etc/apt/sources.list as is will not allow
apt-ostree to install Debian packages via apt.

Also add the following Debian packages because they
are now build depenencies for apt-ostree:

- gir1.2-packagekitglib-1.0
- libpackagekit-glib2-18
- python3-pycurl
- python3-software-properties
- python3-systemd

Test Plan:
1. Build and Install ISO.
2. Check the contents of /etc/apt/sources.list, the file
   should be empty.

Closes-Bug: 2035121

Change-Id: Ic19f786aa62b769058fedd2568692baeeddfee2b
Signed-off-by: Charles Short <charles.short@windriver.com>