debian: /etc/apt/sources.list should be empty
LAT configures the ISO with the apt repositories from
the aptly repositories, which is the correct behaviour.
But it does not clean up after itself and leaves
the /etc/apt/sources.list as is when the build is complete.

Leaving the /etc/apt/sources.list as is will not allow
apt-ostree to install Debian packages via apt.

Test Plan:
1. Build and Install ISO.
2. Check the contents of /etc/apt/sources.list.
3. After patch is applied the contants of /etc/apt/sources.list
   should be empty.

Closes-Bug: 2035121

Change-Id: If3bfcbcfc4e11f42d568602e539d40154778c9bc
Signed-off-by: Charles Short <charles.short@windriver.com>