Add prometheus-client to downloader list
Since the python3-prometheus-client is not build nor downloaded
anywhere, the stx-ceph-manager image is failing to build because of
that. The python3-oslo.metrics (which is a dependency of this image)
reqquires this package.

The purpose of this change is to add this python3-prometheus-client
package to the container downloader list in order to build the
stx-ceph-manager in the flock layer.

This change does not impact the default StarlingX build process as this
is on the `f/antelope` branch.

Test Plan:
PASS: Run downloader for the containers layer
PASS: Build stx-debian image
PASS: Build wheels
PASS: Build stx-ceph-manager image

Closes-Bug: 2035347

Change-Id: Ia404862253b937a8eb27134ad36a28be444d77a6
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>