containers: Fix python-prometheus-client entry
It has been observed that `python-prometheus-client` is failing to
download after being added to the `containers` binaries list [1].

According to snapshot.debian.org [2], the binary for this package
is called `python3-prometheus-client`, and not
`python-prometheus-client`.

Changing the name in the `os-std.lst` file seems to solve the problem.

Note: Since this binary is required for the exclusive build of the
      `stx-ceph-manager` container image, I took the opportunity to try
      to build it after fixing the `downloader`.

      To my surprise, the image build was still failing due to a missing
      package in the `containers` layer: `python3-oslo.messaging`.

      So, since we are already updating this file, I took the
      opportunity to add this missing binary as well.

      (I decided to go with a version compatible with OpenStack
      Victoria, which is what the platform uses)

[1] https://review.opendev.org/c/starlingx/tools/+/894921
[2] https://snapshot.debian.org/package/python-prometheus-client/0.7.1-1.1/

Test Plan:
PASS - Run `downloader`
PASS - Build `stx-debian` base image
PASS - Build `stx-ceph-manager` container image

Closes-Bug: 2036291

Change-Id: I6af526e7fc014228401114baeb2fb0fdce8310a4
Signed-off-by: Luan Nunes Utimura <LuanNunes.Utimura@windriver.com>