Updated minikube profile name
I fixed the issue with minikube_started() and helm_started() where the
grep command was not grepping the correct docker image and helm chart
due to matching substring. It was solved using helm ls command and the
option --kube-context when minikube eixist to pass minikube profile
and helm ls with the option --namespace when minikube is not exist.

TESTS
========================================

* setup a use profile names "MINIKUBENAME" Eg "stx" and "stx1"
* setup a similar environment to reproduce the issue

Closes-Bug: 1995400
Change-Id: Ic9b2224240e0e02be30957a96e7deb24da122136
Signed-off-by: Willy Njomen <willy.njomen@windriver.com>