Remove obsolete stx.conf upgrade
Remove code that automagically changes debian archive URLs in stx.conf
to public WR mirror URLs. This upgrade was added last year and was only
relevant to un-upgraded development environments that existed at that
time (eg on developer desktops). It is no longer relevant. Removing it
because it breaks the ability of the user to explicitly set those URLs
to the real Debian archive repository under some circumstances.

How to reproduce:
* Remove existing stx.conf
* Copy stx.conf.example => stx.conf
* Replace the key:
  stx config --add stx config --add project.debian_snapshot_base http://snapshot.debian.org/archive/debian
* Upgreade configuration:
  stx config --upgrade
* ===> URL gets reset back to default (mirror) in stx.conf

TESTS
==================================
* Re-run above steps and make sure this doesn't happen

Closes-Bug: 2042854

Signed-off-by: Davlet Panech <davlet.panech@windriver.com>
Change-Id: I3ff011d747f2804ade0152ad0296b00ba2785ebf