Add ndisc6 package to iso, allow patches on ifupdown package
The ndisc6 package has useful diagnostic tools for IPv6 networks. It is
being added to allow for duplicate address and gateway reachability
detection by the scripts from the ifupdown-extra package.
The ifupdown package is being removed from the list because it's being
added via the integ project instead, to allow for patches.

Test Plan

[PASS] downloader
[PASS] build-pkgs --clean --all
[PASS] build-image
[PASS] Run full build, system install, bootstrap and unlock SX system
[PASS] Run command "dpkg --list | grep ndisc6"
[PASS] Run command "ndisc6 --help"
[PASS] Run command "dpkg --list | grep ifupdown"
[PASS] Run command "ifup --help"

Depends-On: https://review.opendev.org/c/starlingx/integ/+/908172

Closes-Bug: #2052534
Change-Id: I9dd38bbd1f89e266e0b55ffde9865f94a641c8ff
Signed-off-by: Lucas Ratusznei Fonseca <lucas.ratuszneifonseca@windriver.com>