Changed str comparisons which mixed bash and posix
Executing "source import-stx" produces error message:
"import-stx:175: = not found"

Found that a few string comparisons in the file were using the bash
exclusive string comparison operator '==' within the 'test'
command, which is represented by the single brackets [1].

Basically, string comparisons in single brackets need to use '='.
The proposed syntax is already used in other places in the code,
so this is better for consistency as well.

[1] https://www.man7.org/linux/man-pages/man1/test.1.html

Test Plan:
pass - source import-stx

Change-Id: I4d6bbbaeb9431b6e640de228d123103dd74e0de5
Signed-off-by: Leonardo Fagundes Luz Serrano <Leonardo.FagundesLuzSerrano@windriver.com>