Use the in-tree drivers in initramfs
We had updated the Linux kernel to the latest 6.6.x stable kernel.  As
part of this we want to transition from out-of-tree (OOT) kernel drivers
to in-tree drivers to reduce the development effort going forward and to
benefit from community testing and the latest bug fixes.

In order to deal with a firmware incompatibility or if a problem is
found with the in-tree driver, we want to continue to support one
version of each out-of-tree driver.

That means we will support two kinds drivers in-tree drivers and OOT
drivers in starlingx at the same time.

In the first stage, we will enable and use the in-tree drivers. So
remove OOT drivers from initramfs before enable OOT drivers, or
build-image will fail. We will add OOT drivers back into initramfs
once we support the OOT drivers in the second stage.

Story: 2011056
Task: 49630

Depends-On: https://review.opendev.org/c/starlingx/kernel/+/910323

Change-Id: I9f1027e09cefe297d1a8f8b5154507b5c6dc78ab
Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>