Enable OOT drivers i40e, iavf
In order to deal with a firmware incompatibility or if a problem is
found with the in-tree driver, we want to continue to support one
version of each out-of-tree driver. Since the versions have already
been validated. We will be preserving i40e 2.20.12, iavf 4.5.3.2,
and ice 1.9.11as OOT drivers, along with the NVME update utility
version 9.20 for the 700 series NICs and version 4.22 for the E810
series of NICs.

We only enable i40e and iavf here because ice is already enabled in
the previous commit "Use the in-tree drivers in initramfs".

Verification:
- installs from iso succeed on servers with ice(ntel E810-2C-QDA2
  Chapman beach) and i40e hw(Intel Ethernet Controller X710) for
  rt and std.
- interfaces are up and pass packets for rt and std.
- create vfs, ensure that they are picked up by the new iavf
  driver and that the interface can come up and pass packets
  on rt and std system.
- Check dmesg to see DDP package is loaded successfully and
  the version is 1.3.30.0 for rt and std.

Story: 2011056
Task: 49672

Depends-On: https://review.opendev.org/c/starlingx/kernel/+/922999
Change-Id: I1c4a76ca3e11199349e744e64cc052b50166444e
Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>
(cherry picked from commit 3994bd147d7159e40baf90174b7c85a0ca6672d0)
Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>