Make builder nice to it's host
StarlingX build tools consume excessive cpu and io which can starve
non-build services, including kubernetes and calico. This results in
timeouts, restarts, and general instability of calico and kubernetes
which in turn can the build to fail. This affect is observed on the
main StarlingX build server.

The major build steps should employ 'nice' and 'ionice' to ensure
that critical services on the build host are not starved for cpu or io.

Closes-bug: 2095512
Change-Id: Ie600944989d5d6b760bafa090d25f4891f10943f
Signed-off-by: Scott Little <scott.little@windriver.com>
(cherry picked from commit 6e16428c260fe4de8c69a31b9facff553b1c0051)