Make sbuild nice to it's host
StarlingX build tools consume excessive cpu and io which can starve
non-build services, including kubernetes and calico. This results in
timeouts, restarts, and general instability of calico and kubernetes
which in turn can the build to fail. This affect is observed on the
main StarlingX build server.

The major build steps should employ 'nice' and 'ionice' to ensure
that critical services on the build host are not starved for cpu or io.

Closes-bug: 2095512
Change-Id: I5c96ba200fc5b5cf0dd5dd78b8137535d0330542
Signed-off-by: Scott Little <scott.little@windriver.com>
(cherry picked from commit f360849703b54612f497c3516a96f6cef32799ed)