Upgrade dnspython, eventlet and gunicorn
Upgraded dnspython and eventlet to their bookworm versions
and gunicorn to the lastest version available for Debian.

The core reason is a bug with dnspython 2.0.0, which was preventing
a pkg from being successfully added to an STX patch [1].
Updating it then requires updating eventlet, which then requires
updating gunicorn, because these packages depend on each other
and several bugs appear when they are not compatible with each other.

Refs:
[1] dnspython<=2.0.0 requires /etc/resolv.conf to be available:
https://github.com/eventlet/eventlet/issues/462

Test Plan:
pass: install starlingx-dashboard deb pkg via patch and
      verify system health
pass: build and install ISO and verify system health

Closes-Bug: 2106266

Change-Id: Ic1abf7ea6c8363288880e82a9f5abb80c13259cd
Signed-off-by: Leonardo Fagundes Luz Serrano <Leonardo.FagundesLuzSerrano@windriver.com>