Removing old puppet-module-vswitch package
In addition to the information shared at [1], the puppet-module-vswitch
is now a custom package that was added to the starlingx/integ
repository in version 20. For this reason, it is no longer necessary to
keep the old version of the package in the base-bullseye file.

[1] https://review.opendev.org/c/starlingx/integ/+/951436

TEST PLAN:

PASS - downloader --clean
PASS - build-pkgs -c -p puppet-module-vswitch,puppet-manifests
PASS - build-pkgs --all
PASS - build-image
PASS - AIO-SX fresh install with OVS as vswitch type

Depends-On: https://review.opendev.org/c/starlingx/integ/+/951436

Story: 2011263
Task: 52281

Change-Id: I974cc885f361cc4b0bb6e1cb7686322ec7dd289e
Signed-off-by: Murillo Arantes <murillo.arantes@windriver.com>