Debian: librabbitmq-dev: change version
Upgrade librabbitmq-dev to 0.10.0-1+deb11u1

The current librabbitmq-dev package has a dependency on
librabbitmq4 version 0.10.0-1. However, due to the change
made in [1], it is also necessary to update the
librabbitmq-dev version package to the same version of
librabbitmq4 that is 0.10.0-1+deb11u1.

[1]: https://review.opendev.org/c/starlingx/tools/+/950319

TestPlan:
PASS: downloader; build-pkgs
PASS: build-image
PASS: install on SX-lab (VBox)

Closes-Bug: 2105457

Change-Id: I2f83440bb3b09a8ae7a3a63fbe083f047b8a520f
Signed-off-by: Joao Tognolli Jr <Joao.TognolliJunior@windriver.com>