Fixes to allow trixie to build within trixie containers
For the purposes of the downloader, the apt was reusing the
/etc/apt/sources.list of the host container, rather than
specifying a seperate sources lists for builder container
and downloader.  This means you can only build trixie in a
trixie build container.   Not a terrible assumption, but there
is tension.  Our lst files specify specific versions of packages.
In order to ensure those versions are available, we need
/etc/apt/sources.list to point to a trixie snapshot, and not
to the patch current trixie repos.   So the build containers
can't benefit from bug fixes, or the /etc/apt/sources.list needs
to be hacked after the build container  is created, but before
a build is attempted.  Hacks that may have to be undone if your
trying to fix issues in the build container.

I think it's desirable to seperate the sources.list of the
build container from the sources.list used by stx downloader.
THis update only goes part way to solving this.

1) Switch the sources.list of the build container to use a trixie
snapshot.  Hopefully this is a temporary measure. (see 3)

2) Create a seperate apt environment in a chrootable path for
use by the downloader

3) Copy the build container sources.list into the chroot.
This should be a temporary measure, until a better method to
populate it, independent if the host, is worked out.

4) Modify downloader to use the alternate apt environment.
For most commands, it's just a matter of pointing apt at
an alternate apt config file.  However 'apt update' seems to
be an exception, requiring entry into a chroot to keep it
independent of the hosts lock files and cache. This in turn
means we need two versions of the alternate apt config,
featuring slightly different paths...
One for when we are within the chroot (apt update), and one for
when we are not in the chroot (apr query).

Next, the trixie version of flask fixed a longstanding bug that
blurred the lines between 'GET' and 'POST' transactions.
The starlingx build tools were sending and recieving 'GET',
but interpreting the recieved data using 'POST' style syntax.
The buggy old flask allowed us to get away with this.
No so the new flask.

Also there was a lot of issues with error handling on both
server and client side of these transaction that made debugging difficult.

5) Continue to use 'GET', but arguements must now be accessed
with request.args, not request.form.

6) Fix how errors are handled when a request fails.  seperate errors for
400 - server not ready, and 500 - internal error, and internal errors are
logged

Finally there is stx/stx-build-tools-chart/stx-builder/configmap/stx-sources.list.
This file is bullseye specific.

7) rename it with a bullseye qualified name and create a trixie version as well.

Note, these files are not referenced by default, they are merely copied to /usr/local/bin/stx
in the builder container.  A 'sources_list = <path>' directive would have to be added
to the stx.conf to make use of this file.

8) Prepare a chroot within the builder for apt queries

9) Allow parallel apt downloads

Story: 2011360
Task: 51933
Change-Id: I251e8f1481e5418c4ef219ce0c8cb12914ffd2a1
Signed-off-by: Scott Little <scott.little@windriver.com>