Debian distro env var in lat container
When creating the build environment, the dev has to define
the Debian Distribution which will be used as base for the
STX ISO (either 'bullseye' or 'trixie' at the moment).

With this change, this value will be available inside
the stx-lat-tool build container as the DIST env variable.

Test Plan:
pass: DIST available in 'stx shell --container lat'

Story: 2011498
Task: 52947

Change-Id: I98d0a19650eed61e41391f0b6aa9679daaa4dd18
Signed-off-by: Leonardo Fagundes Luz Serrano <Leonardo.FagundesLuzSerrano@windriver.com>