stx-builder: avoid broken docker client
Latest (as of now) docker client, 5:28.5.2-1~debian.11~bullseye, is
broken and can't login to docker.io. The login error we get is:

  client version 1.52 is too new. Maximum supported API version is 1.41

There's some evidence on the Internet [1] of similar errors happenning in
the past due to a version comparison bug in docker client code.

Workaround: install the last known to be working version in builder Dockerfile.

We should remove this workaround once Debian deploys a version that
doesn't have this issue.

TESTS
==============
* Build "stx-builder" container & start the environment
* docker login -u USER
==> Login fails
* Downgrade "docker-ce-cli" package to 5:28.5.2-1~debian.11~bullseye
* docker login -u USER
==> Login succeeds

[1] https://stackoverflow.com/questions/62079794/error-response-from-daemon-client-version-1-40-is-too-new-maximum-supported-ap

Signed-off-by: Davlet Panech <davlet.panech@windriver.com>
Change-Id: Ia3199e864c65c50a0769bcc93592fbc1caee2968