Update python-ironicclient dependency
As an effort for the STX Openstack upversion to the Epoxy
release, some packages need to be updated, such as the
python-ironicclient.

This upversion (5.5.0-3 -> 5.10.0-4) required a newer version
of the python-pbr package. [1]

[1] - https://salsa.debian.org/openstack-team/clients/python-ironicclient/-/compare/debian%2F5.5.0-3..debian%2F5.10.0-4?from_project_id=574

Test Plan:
- PASS: Build all STX-O packages and wheels
- PASS: Build STX-O tarball
- PASS: Build STX docker images
- PASS: Build all STX packages and STX ISO
- PASS: STX ISO in a virtual machine
- BLOCKED: Apply STX-O overriding generated images

The apply failed due to an issue related to db_sync, which
will be  handled in the Task: 53501

Story: 2011516
Task: 52973

Change-Id: I9820c091ff685a03efe1aadf27856fa83deb7d4f
Signed-off-by: jbarboza <joaopedro.barbozalioneza@windriver.com>